auth/requirements.txt
Django
djangorestframework
djoser
djangorestframework_simplejwt
mysqlclient
django-cors-headers
pika
python-dotenv
psycopg2